rename: 아이콘 경로 수정
